backend/requirements.txt
django==5.1.4
djangorestframework
# environs
django-environ
psycopg2-binary
djangorestframework-simplejwt
dj-rest-auth
django-allauth
django-cors-headers
django-cities-light
"django-phonenumber-field[phonenumbers]"
django-rest-swagger
drf-yasg
Pillow
django-markdownx